refactor: HashtagName을 별도 파일로 분리
- `BottariTemplateHashtag.kt`에 정의되어 있던 `HashtagName` 값 클래스와 관련 오류(`HashtagNameError`)를 `HashtagName.kt` 파일로 분리하여 관리하도록 구조를 개선합니다.
